# docs/swagger.yaml
basePath: /api
definitions:
...
  req.RequestCreateHotel:
    properties:
      activate:
        type: boolean
      apartment:
        type: boolean
      bankAccount:
        type: string
      bankBeneficiary:
        type: string
      bankName:
        type: string
      bar:
        type: boolean
      bath:
        type: boolean
      beach:
        type: boolean
      breakfast:
        type: boolean
      bussiness_license:
        type: string
      camping:
        type: boolean
      casio:
        type: boolean
      district:
        type: integer
      fitness:
        type: boolean
      homestay:
        type: boolean
      hotel:
        type: boolean
      hotelier_id:
        type: string
      id:
        type: string
      motel:
        type: boolean
      name:
        type: string
      no_smoking_room:
        type: boolean
      overview:
        type: string
      parking:
        type: boolean
      photos:
        type: string
      pool:
        type: boolean
      province:
        type: integer
      raw_address:
        type: string
      resort:
        type: boolean
      spa:
        type: boolean
      villa:
        type: boolean
      ward:
        type: integer
      wifi:
        type: boolean
    type: object
...
  req.RequestUpdateHotel:
    properties:
      activate:
        type: boolean
      apartment:
        type: boolean
      bankAccount:
        type: string
      bankBeneficiary:
        type: string
      bankName:
        type: string
      bar:
        type: boolean
      bath:
        type: boolean
      beach:
        type: boolean
      breakfast:
        type: boolean
      camping:
        type: boolean
      casio:
        type: boolean
      district:
        type: integer
      fitness:
        type: boolean
      homestay:
        type: boolean
      hotel:
        type: boolean
      motel:
        type: boolean
      name:
        type: string
      no_smoking_room:
        type: boolean
      overview:
        type: string
      parking:
        type: boolean
      pool:
        type: boolean
      province:
        type: integer
      raw_address:
        type: string
      resort:
        type: boolean
      spa:
        type: boolean
      villa:
        type: boolean
      ward:
        type: integer
      wifi:
        type: boolean
    type: object
...
  req.RequestUpdateProfile:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
...
  res.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: https://siphoria-api-production.up.railway.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HOTEL BOOKING API
  version: "1.0"
paths:
...
  /admin/push-notification:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin push notifications
      tags:
      - admin-service
...
  /hotel/:id:
    patch:
      consumes:
      - application/json
      parameters:
      - description: hotel
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateHotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: update hotel info
      tags:
      - hotel-service
  /hotels:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get hotel Controller
      tags:
      - hotel-service
    post:
      consumes:
      - application/json
      parameters:
      - description: hotel
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestCreateHotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Create hotel
      tags:
      - hotel-service
  /hotels/:id:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get hotel by Id
      tags:
      - hotel-service
  /hotels/:id/business-license:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update hotel business license photos
      tags:
      - hotel-service
  /hotels/:id/payout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Send request payment hotel
      tags:
      - hotel-service
  /hotels/:id/photos:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Delete hotel photos
      tags:
      - hotel-service
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update hotel photos
      tags:
      - hotel-service
  /hotels/:id/rooms:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Get room type by hotel
      tags:
      - hotel-service
  /hotels/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: hotel
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Search Hotel
      tags:
      - hotel-service
...
swagger: "2.0"